+++ RobWork/src/rwlibs/calibration/ParallelAxisDHJacobian.cpp
@@ -71,7 +71,7 @@ namespace rwlibs {
 					columnIndex++;
 				}
 
-				RW_ASSERT(columnIndex == columnCount);
+				RW_ASSERT(columnIndex == (int)columnCount);
 
 				return jacobian;
 		}
