+++ RobWorkSim/src/rwsimlibs/ode/ODEKinematicDevice.cpp
@@ -72,7 +72,7 @@ namespace {
 
     Jacobian subJacobian(const Jacobian& jac, int toCol){
         int m = (int)jac.size1();
-        RW_ASSERT(toCol<=jac.size2());
+        RW_ASSERT(toCol<=(int)jac.size2());
         Jacobian jsub(m,toCol);
         for(int x=0;x<m;x++)
             for(int y=0;y<m;y++)
@@ -91,7 +91,7 @@ namespace {
     {
         using namespace boost::numeric::ublas;
         int m = (int)Jq.size1();
-        RW_ASSERT(column<Jq.size2());
+        RW_ASSERT(column<(int)Jq.size2());
 
         matrix<double> subjac(m,column+1);
         matrix_range<matrix<double> > mr (subjac, range (0, m), range (0, column+1));
